backend/requirements.txt
# ===============================
# CORE BACKEND
# ===============================
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.35
psycopg[binary]==3.1.19
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
aiofiles==24.1.0


# ===============================
# FILE PROCESSING
# ===============================
PyPDF2==3.0.1
python-docx==1.1.2
pypdf==4.2.0
docx2txt==0.8


# ===============================
# LLAMAINDEX (FINAL WORKING SET)
# ===============================
llama-index==0.10.43
llama-index-core==0.10.43
llama-index-embeddings-openai==0.1.5
llama-index-llms-openai==0.1.14
llama-index-vector-stores-pinecone==0.1.4
pinecone-client==3.2.2

supabase==2.5.0


# ===============================
# LLM PROVIDERS
# ===============================
openai==1.56.0
anthropic==0.34.2


# ===============================
# TOKENIZATION
# ===============================
tiktoken==0.7.0


# ===============================
# EXPORT / REPORT GEN
# ===============================
reportlab==4.2.2
python-pptx==0.6.23
weasyprint==62.3


# ===============================
# LANGCHAIN + LANGGRAPH
# ===============================
langgraph==0.2.28
langchain==0.3.7
langchain-openai==0.2.5
langchain-community==0.3.5
langsmith==0.1.129
